Unroll loop in evaluator and hacks to Sobel
Ignore-this: 5fce8491c293a50d13372249ec02a366

darcs-hash:20110216035412-bf82d-b502f160156ef4541a2a6ddf12a3a6b06ba9a34d

+++ repa-examples/examples/Sobel/src-repa/Solver.hs
@@ -3,12 +3,14 @@
 
 module Solver 
 	( Image
-	, gradientX, gradientY )
+	, gradientX
+	, gradientY )
 where
 import Data.Array.Repa 			as Repa
 import Data.Array.Repa.Stencil
 
-type Image	= Array DIM2 Double
+type Image	= Array DIM2 Float
+
 
 gradientX :: Image -> Image
 {-# NOINLINE gradientX #-}
@@ -28,4 +30,6 @@ gradientY img@Manifest{}
 	  [stencil2|	 1  2  1
 			 0  0  0
 			-1 -2 -1 |] 
+
+
